Improvements to debug logging by implementing a custom alternative to NSLog and adding nice logging functions to the webserver, websocket and the web library. Some debug output about iBeacon detection was removed to reduce log flooding.
This commit introduces a first draft at the information flow between native layer, webserver and the web library. The web library is now informed about the local device information as well as new and changed ibeacons nearby. Furthermore, the native layer is informed when the local websocket connection has been successfully established.

We introduce LazyLoad in order to load the different files that make up the web library easily. LazyLoad has to be loaded in the head of the web app before connichiwa.

Lastly, some minor improvements like refined names, parameters and exceptions in the weblib.