Removed native DeviceManager
Great, job, eh? Seriously, I thought long and hard about how to implement this nicely, and the nativemanager just didn’t seem to be the way to go. It didn’t add anything useful and caused code duplications and weird communication in the native app.
Therefore, the functionality of the devicemanager has now directly been incorporated into the BeaconMonitor. It keeps track of new and changed beacons and sends messages to the delegate accordingly, which sends messages to the weblibrary.
Furthermore, the singleton status of most classes was revoked - singletons and delegates don’t go well together, and I didn’t want to have to use NSNotificationCenter, as it can be easily misused to code dirty.
The weblibrary picks up the new messages, but its not completely implemented yet.
Furthermore, I tried to change some minor things to make stuff more clear, for example we don’t send a local device anymore but just a local id.
Also, we keep everything related to CoreLoacation in the beacon classes. CWWebApplication, CWWebserver and others should not be concerned with CoreLocation.