First Draft of Direct WebView Communication
Via JavaScriptCore we can directly communicate with the remote library and the web library - this commit has a first draft on exactly that.
It introduces CWWebLibraryManager and CWRemoteLibraryManager that load the libraries on command, receive and delegate commands from it and can send messages to it.

There is still lots of old and unfitting code in there, though, so there is still a lot to do.