New Debug System
Introducing the new debug system with sources and priority
Every log message from all sources is now gathered in XCode’s output and has a priority and a source assigned to it. A higher priority means the message is more likely to spam the log output, but can give deeper insights. Debug messages can be filtered by both priority and source.