The local device information are now complete
The way the native layer sends the local device information to the JS library was changed. The local device information are now identical to the initial data sent to other devices via BT. For that, the local device information dictionary is now built by the AppState and used by both CWWebLibraryManager and CWBluetoothManager. Furthermore, the way these information are received have been changed: The master web library now passes this dictionary to CWDeviceManager which creates the local device with all the information. In turn, Connichiwa’s _setIdentifier has been taken out of the master library and getIdentifier now forwards to the getIdentifier of the local device.