constructor support in OOP & minor changes to debug
OOP.js now supports singleton constructors by using the magic method name __constructor. If such a method exists, it is now automatically called after the singleton was built. Note that if a singleton is extended, right now the constructor is called again. This might change in the future.

The debug message is now sent to the webview earlier and Connichiwa now builds without the CWDEBUG flag being set