Master devices can now be stitched and unstitched like every other device
Previously, the first stitch needed to include the master device - the master device was always the first device that was stitched. Now, every two devices can participate in the first stitch. Furthermore, the master device now behaves like every other device and can also be unstitched and restitched as necessary. The global coordinate system will now be the coordinate system of the first stitched device.

- wasStitched and wasUnstitched were moved into common because master and remotes need them now
- CWStitchManager now supports an .unstitch() method which will unstitch the current device manually
- CWLocation now automatically adjusts itself when a device is stitched/unstitched. This means that all existing CWLocation's will represent the same LOCAL location after a stitch/unstitch. So if a CWLocation represents a DOM element, the DOM element will not be altered by a stitch/unstitch.