Fix #81
A native-less remote will now try to reconnect the websocket when it gets disconnected. Since it seems impossible to restore a remotes state (because it can depend on so many things like loaded scripts, loaded DOM objects, ...) as soon as the websocket has been reestablished we reload the page, which will reset the remote and reestablish a connection.