Master is now preferred on the first stitch
If the master is involved in the very first stitch it is now preferred (stitched first), making the global coordinate system the same as the master's coordinate system. Most applications will not care about that, but some might rely on the master being stitched first, for example because they start with content on the master only and don't want that content to change on the first stitch.