Fix #89
Added Connichiwa.onLoad() (or, alternatively, Connichiwa.on('load', ...)). Callbacks attached to this will fire on $(window).load() or immediatly when the window is already load. This can be important if we have common code: On the master, we want it to fire on $(window).load(), but since it is loaded via loadScript() on remotes, it will arrive AFTER the window has loaded, and therefore needs to fire immediately. Connichiwa.onLoad() takes care of that and should be preferred over $(window).load(), window.onload or body.onload.