Further improved Websocket server
We now get notified immediately when a websocket event occurs, instead of polling for it every few milliseconds. This is done by using libwebsocket_service() with a large timeout, which will block until an event occurs and then immediately call the callback. SInce we do this on a seperate thread, the blocking behaviour is not a problem. Further, we changed the dispatch timer's so it fires instantly, which means that if an event occurs and libwebsocket_service() stops blocking, libwebsocket_service() will be called immediately after the event was handled and waits for new events.