Improved CWModule
CWModules.add() was removed. Instead, modules should now use CWModules.retrieve() to get back a module object. The module object is an object literal, but CWModules.init() will perform some magic on it: It will call its __constructor() and will make sure that "this" is available in all module methods.