Added master backlog to websocket server
It was possible that a quickly loading remote could send its remoteinfo message before the master websocket connection was established. The remote would then not be recognized. This happened particularly on auto-reload. To solve it, we re-introduce (-_-) the master backlog - messages that arrive for the master before its connection is ready are remembered and delivered when the master websocket ist established.